pyproject.toml
[build-system]
requires = ["setuptools>=64", "setuptools-scm>=8"]
build-backend = "setuptools.build_meta"

[project]
name = "photomesh"
version = "0.1.0"
description = "Texture mapping between camera frames and 3D meshes using camera poses"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "PhotoMesh Contributors"},
]
keywords = ["3d", "texture", "mesh", "reconstruction", "camera", "uv-mapping"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Image Processing",
    "Topic :: Multimedia :: Graphics :: 3D Modeling",
]
dependencies = [
    "numpy",
    "open3d",
    "Pillow",
    "opencv-python-headless",
    "xatlas",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff",
]

[project.urls]
Homepage = "https://github.com/photomesh/photomesh"
Repository = "https://github.com/photomesh/photomesh"

[tool.setuptools.packages.find]
include = ["photomesh*"]
